+++ src/mupdfjs.ts
@@ -20,8 +20,8 @@
 // Artifex Software, Inc., 39 Mesa Street, Suite 108A, San Francisco,
 // CA 94129, USA, for further information.
 
-import * as mupdf from "mupdf";
-export * from "mupdf";
+import * as mupdf from "./core.js";
+export * from "./core.js";
 
 export type PDFWord = {
 	rect: mupdf.Rect,
